refactor(kv): move kv to a standalone package to prevent loading sqlite at startup
