+++ plugins/manager.go
@@ -7,6 +7,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/emqx/kuiper/common"
+	"github.com/emqx/kuiper/common/kv"
 	"github.com/emqx/kuiper/xstream/api"
 	"io"
 	"io/ioutil"
@@ -281,7 +282,7 @@ type Manager struct {
 	pluginDir string
 	etcDir    string
 	registry  *Registry
-	db        common.KeyValue
+	db        kv.KeyValue
 }
 
 func NewPluginManager() (*Manager, error) {
@@ -302,7 +303,7 @@ func NewPluginManager() (*Manager, error) {
 			outerErr = fmt.Errorf("cannot find db folder: %s", err)
 			return
 		}
-		db := common.GetSqliteKVStore(path.Join(dbDir, "pluginFuncs"))
+		db := kv.GetDefaultKVStore(path.Join(dbDir, "pluginFuncs"))
 		err = db.Open()
 		if err != nil {
 			outerErr = fmt.Errorf("error when opening db: %v.", err)
